# Fix NLog integration when using ILogger