@pjanotti 2021-12-08
Test failures seem unrelated, kicking it again. The following test seems flaky on Windows: `Datadog.Trace.ClrProfiler.IntegrationTests.AdoNet.SqlCommandTests.SubmitsTraces(packageVersion: "")`